Allow user to make selection via button click
Rather than displaying a prompt for the user to type in their choice,
which could cause frustration if the wrong string was entered, this
commit replaces the prompt with three buttons that correspond to the
available choices.

Clicking a button plays a round of the game with the user's choice being
the button that they clicked.